WIP: Clean up old code
