--- learn/models/__init__.py
@@ -1,1 +1,0 @@
-from .infogan import InfoGAN
